test: deprocess maintenance API fixtures
